Extract pagination logic
