+++ backend/internal/post/post.go
@@ -2,9 +2,9 @@ package post
 
 import (
 	"github.com/doktorupnos/crow/backend/internal/model"
+	"github.com/doktorupnos/crow/backend/internal/pages"
 	"github.com/doktorupnos/crow/backend/internal/user"
 	"github.com/google/uuid"
-	"gorm.io/gorm"
 )
 
 type Post struct {
@@ -22,7 +22,7 @@ type FeedPost struct {
 
 type LoadParams struct {
 	Order string
-	PaginationParams
+	pages.PaginationParams
 }
 
 type PostRepo interface {
@@ -31,16 +31,4 @@ type PostRepo interface {
 	LoadByID(id uuid.UUID) (Post, error)
 	Update(p Post) error
 	Delete(p Post) error
-}
-
-type PaginationParams struct {
-	PageNumber int
-	PageSize   int
-}
-
-func Paginate(params PaginationParams) func(db *gorm.DB) *gorm.DB {
-	return func(db *gorm.DB) *gorm.DB {
-		offset := (params.PageNumber - 1) * params.PageSize
-		return db.Offset(offset).Limit(params.PageSize)
-	}
 }
